+++ backend/scoreMaker.ts
@@ -253,12 +253,7 @@ const makeScore = async (
 		logger.append("makeScore.match", {coverage, omitC, omitS, path: matcher.path});
 	}
 
-	const midiNotation = lilyNotation.toPerformingNotation();
-
-	const matchedIds: Set<string> = new Set();
-	midiNotation.notes.forEach(note => note.ids && note.ids.forEach(id => matchedIds.add(id)));
-
-	doc.updateMatchedTokens(matchedIds);
+	doc.updateMatchedTokens(lilyNotation.idSet);
 
 	if (baking)
 		doc.pruneForBakingMode();